Move the API error handling into one GrapeErrors file
Three files across app/controllers, app/services and lib for one subsystem was over-split,
and none of it was a Rails controller. GrapeErrors holds the exception, the status table,
the message scrubbing, the report? predicate, the logging/notify, and the Rack middleware -
Zeitwerk only constrains the top-level constant for a path, so nesting is free. Named for
Grape rather than API:: to keep it clear of APIAuthorization::Errors.

Restores format/default_error_formatter/content_type to both roots. Hoisting them onto
API::Base relied on Grape's inheritance for mounted subclasses, and default_error_formatter
governs the error! throw path, which no spec reaches - not worth six lines.
